traildb_bench: new commands info, dump, recode
most important change is command 'recode':

= recode =

recode lets you create a new DB with a subset of
the fields. The idea is to combine this with
another script from trick that determines the
list of required fields such that minimal,
faster to process TrailDB instances can be
created.

= info =

dumps TrailDB statistics including set of
existing fields.

= dump =

dumps hex-encoded content of a TrailDB to stdout