Fix double-close of filedescriptor.
cons_package() was always calling close() twice on the same file-descriptor.
This could create problems if the filedescriptor had already been allocated
to another file in another thread. We also now avoid to call close() twice
in case of errors, as POSIX says that the fd is always invalid after close(),
regardless of the error code.

Moreover, also archive_write_free() was being called twice in case of
errors. This is also probably a bug as it could trigger double-free of
heap-allocated memory.